#9: Removed code duplication in database object that are schema children.
