+++ src/main/java/com/exasol/dbbuilder/AdapterScript.java
@@ -5,8 +5,7 @@ import com.exasol.dbbuilder.objectwriter.DatabaseObjectWriter;
 /**
  * Virtual Schema Adapter Script
  */
-public class AdapterScript extends AbstractDatabaseObject {
-    private final Schema parentSchema;
+public class AdapterScript extends AbstractSchemaChild {
     private final Language language;
     private final String content;
 
@@ -21,8 +20,7 @@ public class AdapterScript extends AbstractDatabaseObject {
      */
     public AdapterScript(final DatabaseObjectWriter writer, final Schema parentSchema, final String name,
             final Language language, final String content) {
-        super(writer, name);
-        this.parentSchema = parentSchema;
+        super(writer, parentSchema, name);
         this.language = language;
         this.content = content;
         writer.write(this);
@@ -31,16 +29,6 @@ public class AdapterScript extends AbstractDatabaseObject {
     @Override
     public String getType() {
         return "adapter script";
-    }
-
-    @Override
-    public boolean hasParent() {
-        return true;
-    }
-
-    @Override
-    public DatabaseObject getParent() {
-        return this.parentSchema;
     }
 
     /**
